Read full commit
* Make a custom printf and put the old printf under `cprintf`
* Add the string::format function that works similar to rust's format!() macro (doesnt support template literals)
* Don't propagate the maintained type to function call parameters;

If you have a function `foo(string a)`  and call it like `foo(string::format("{}", 1))` this will cause `1` to be casted into a string because the `string a` propagates into the parameter of format().

This is no longer the case, and is definitely unwanted behavior. Thankfully this change does not break any assumptions made in the standard library or examples.