fix: regression in parser
Specifically in cases when you wrap walrus operator expressions `x := y` in parenthesis and then use that as part of a bigger arithmetic operation, the parser interprets the `:` as the ternary end token and decrements the nesting depth.

An example is `if (x := 1) < 2 {}` which failed before but works now. You can read more about this in the associated test: expr_assign.le