fix: also use the current dir when importing
beforehand, if there was a module `./foo.le` and a dir `lib` with `bar.le` and `foo.le` inside, and then you `use foo` in `main.le` but then `use foo` in `bar.le`, the compiler would determine that the module was already imported even though it wasn't, because the path is different those are 2 different modules. this fixes that by also using the PathBuf of the dir which the module will be imported from in the `parsed_modules` HashSet.